backend/requirements.txt
# PDF reading
Pillow~=10.1.0
PyMuPDF==1.23.5
pyzbar~=0.1.9

# Selenium
selenium~=4.15.2
webdriver-manager~=4.0.1

# verify-code preprocess
opencv-python~=4.8.1.78
numpy~=1.24.4
requests~=2.31.0
clint~=0.5.1

# export excel
openpyxl~=3.1.2
python-docx

# flask backend
flask
